+++ src/main/java/utility/ShowView.java
@@ -23,13 +23,13 @@ public class ShowView {
 			parent.thirdDimension = parent.thirdDimensionSize;
 			
 			
-			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg, parent.thirdDimension,
-					parent.thirdDimensionSize, parent.fourthDimension, parent.fourthDimensionSize);
+			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg, (int)parent.thirdDimension,
+					(int)parent.thirdDimensionSize, (int)parent.fourthDimension, (int)parent.fourthDimensionSize);
 			
 		} else {
 
-			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg, parent.thirdDimension,
-					parent.thirdDimensionSize, parent.fourthDimension, parent.fourthDimensionSize);
+			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg, (int)parent.thirdDimension,
+					(int)parent.thirdDimensionSize, (int)parent.fourthDimension, (int)parent.fourthDimensionSize);
 			
 		}
 
@@ -45,13 +45,13 @@ public class ShowView {
 			parent.fourthDimension = parent.fourthDimensionSize;
 			
 			
-			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg, parent.thirdDimension,
-					parent.thirdDimensionSize, parent.fourthDimension, parent.fourthDimensionSize);
+			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg,(int) parent.thirdDimension,
+					(int)parent.thirdDimensionSize,(int) parent.fourthDimension, (int)parent.fourthDimensionSize);
 			
 		} else {
 
-			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg, parent.thirdDimension,
-					parent.thirdDimensionSize, parent.fourthDimension, parent.fourthDimensionSize);
+			parent.CurrentView = utility.Slicer.getCurrentView(parent.originalimg,(int) parent.thirdDimension,
+					(int)parent.thirdDimensionSize, (int)parent.fourthDimension, (int)parent.fourthDimensionSize);
 			
 		}
